Create view controller
